feat: Payment
Done Payment, but not yet checking and callback to validate the payment

+++ database/migrations/2025_10_27_084806_create_transaction_log.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('user_id')->nullable()->constrained('users')->nullOnDelete();
             $table->foreignId('donation_id')->nullable()->constrained('donations')->nullOnDelete();
             $table->string('status')->default('pending');
-            $table->decimal('gross_amount', 15, 2);
+            $table->bigInteger('gross_amount');
             $table->boolean('is_anon')->default(false);
             $table->timestamps();
         });
